Add empty output.log handling

+++ server.sh
@@ -5,7 +5,9 @@ AS_OUTPUT="/var/www/html/alpha-server/report.txt"
 
 while true; do
 	while IFS=, read SUM HOSTNAME; do
-		echo "$HOSTNAME had $SUM attempts." > $AS_OUTPUT
+		if [ -s $AS_INPUT ]; then
+			echo "$HOSTNAME had $SUM attempts." > $AS_OUTPUT
+		fi
 	done <<< $(cut -d, -f2 $AS_INPUT | sort | uniq -c | sed 's/ *//' | sed -r 's/\s+/,/')
 	sleep 5
 done
